# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Sarbon API — документация и руководство
  version: 0.1.0
  description: |
    **Sarbon** — API для логистики и грузоперевозок: водители, диспетчеры, компании, грузы, офферы, рейсы. Полное руководство по API на русском языке: потоки (flow), логика, описание полей и примеры.

    **Как пользоваться Swagger:**
    1. Нажмите **Authorize** и укажите заголовки: **X-Device-Type** (web / ios / android), **X-Language** (ru / uz / en / tr / zh), **X-Client-Token**. Они обязательны для всех запросов к /v1 и /api.
    2. Для методов, требующих авторизации, дополнительно укажите **X-User-Token** (JWT access_token после входа водителя, диспетчера или админа).
    3. Вкладки (теги) сверху группируют методы по ролям: водители, диспетчеры, админ, грузы (Cargo), чат, компании, справочники (Reference).

    **Категории API:** Водители (Drivers) — регистрация, профиль, KYC, рейсы; Фриланс-диспетчеры — вход, профиль, создание грузов, приглашения водителей, назначение на рейсы; Админ — компании и владельцы; Грузы (Cargo) — создание, список, офферы, принятие оффера; Справочники (Reference) — города, типы кузова, статусы.
    В каждом разделе вверху дано краткое описание потока; у схем и методов — подробные русские пояснения к полям.

    **Полная документация на русском:** см. файл docs/DOCUMENTATION-RU.md в репозитории.
servers:
  - url: /
security:
  - DeviceTypeHeader: []
    LanguageHeader: []
    ClientTokenHeader: []
...
components:
  securitySchemes:
    DeviceTypeHeader:
      type: apiKey
      in: header
      name: X-Device-Type
      description: 'Обязателен для всех запросов. Допустимые значения: web, ios, android. Один раз выбранное значение действует для всех разделов (Drivers, Freelance Dispatchers, Admin, Cargo, Chat, Company).'
    LanguageHeader:
      type: apiKey
      in: header
      name: X-Language
      description: 'Обязателен для всех запросов. Допустимые значения: ru, uz, en, tr, zh. Один раз выбранное значение действует для всех разделов (любой язык в любом разделе).'
    ClientTokenHeader:
      type: apiKey
      in: header
      name: X-Client-Token
      description: Обязателен для всех API-запросов (включая Admin). Проверяется сервером на каждом запросе /v1/*.
    UserTokenHeader:
      type: apiKey
      in: header
      name: X-User-Token
      description: JWT access token. Для разделов Admin — обязательно JWT с role=admin (после /v1/admin/auth/login/password); сервер отклоняет запрос при другой роли или неверном токене.
    UserIDHeader:
      type: apiKey
      in: header
      name: X-User-ID
      description: 'Для теста чата в Swagger — UUID пользователя (водитель/диспетчер). Подставляется вместо JWT.'
  schemas:
    Envelope:
      type: object
      properties:
        status: { type: string, enum: ["success","error"] }
        code: { type: integer, example: 200 }
        description: { type: string, example: "ok" }
        data: { nullable: true }
      required: [status, code, description, data]
...
    ItemWithLabel:
      type: object
      description: Элемент справочника (value — код для API, label — подпись для UI).
      properties:
        value: { type: string }
        label: { type: string }
...
    RoleRef:
      type: object
      description: Роль в компании (из app_roles).
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
    ReferenceDriversResponse:
      type: object
      properties:
        registration_step: { type: array, items: { $ref: "#/components/schemas/ItemWithLabel" } }
        registration_status: { type: array, items: { $ref: "#/components/schemas/ItemWithLabel" } }
        driver_type: { type: array, items: { $ref: "#/components/schemas/ItemWithLabel" } }
        work_status: { type: array, items: { $ref: "#/components/schemas/ItemWithLabel" } }
        power_plate_types: { type: array, items: { $ref: "#/components/schemas/ItemWithLabel" } }
        trailer_plate_types_by_power: { type: object, additionalProperties: { type: array, items: { $ref: "#/components/schemas/ItemWithLabel" } } }
...
    ReferenceCompanyResponse:
      type: object
      properties:
        company_type: { type: array, items: { $ref: "#/components/schemas/ItemWithLabel" }, description: "Только SHIPPER, CARRIER, BROKER (верхний регистр)" }
        company_status: { type: array, items: { $ref: "#/components/schemas/ItemWithLabel" }, description: "ACTIVE, INACTIVE, BLOCKED, PENDING" }
        company_user_roles: { type: array, items: { $ref: "#/components/schemas/ItemWithLabel" }, description: "Роли пользователей компании: OWNER, CEO, TOP_MANAGER, TOP_DISPATCHER, DISPATCHER, MANAGER" }
        roles: { type: array, items: { $ref: "#/components/schemas/RoleRef" }, description: "Роли из БД (id, name) для приглашений" }
    ReferenceAdminResponse:
      type: object
      properties:
        admin_status: { type: array, items: { $ref: "#/components/schemas/ItemWithLabel" } }
        admin_type: { type: array, items: { $ref: "#/components/schemas/ItemWithLabel" } }
    ReferenceDispatchersResponse:
      type: object
      properties:
        work_status: { type: array, items: { $ref: "#/components/schemas/ItemWithLabel" } }
...
paths:
...
  /v1/reference/drivers:
    get:
      tags: ["Reference / Drivers"]
      summary: "Справочник Drivers — статусы, типы, шаги регистрации"
      description: "Все допустимые значения для раздела водителей. Фронт и мобилка могут вызывать один раз и кэшировать."
      security: [{ DeviceTypeHeader: [] }, { LanguageHeader: [] }, { ClientTokenHeader: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/ReferenceDriversResponse" }
  /v1/reference/cargo:
    get:
      tags: ["Reference / Cargo"]
      summary: "Справочник Cargo — статусы груза, точки, офферы, типы ТС"
      description: "Все допустимые значения для раздела грузов."
      security: [{ DeviceTypeHeader: [] }, { LanguageHeader: [] }, { ClientTokenHeader: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/ReferenceCargoResponse" }
  /v1/reference/company:
    get:
      tags: ["Reference / Company"]
      summary: "Справочник Company — типы компании, статусы, роли пользователей"
      description: "company_type — только CARGO_OWNER, CARRIER, EXPEDITOR. company_user_roles — только OWNER, CEO, TOP_MANAGER, TOP_DISPATCHER, DISPATCHER, MANAGER. company_status — ACTIVE, INACTIVE, BLOCKED, PENDING. roles — из БД (id, name, description) для приглашений. Все value в верхнем регистре."
      security: [{ DeviceTypeHeader: [] }, { LanguageHeader: [] }, { ClientTokenHeader: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/ReferenceCompanyResponse" }
  /v1/reference/admin:
    get:
      tags: ["Reference / Admin"]
      summary: "Справочник Admin — статусы и типы админа"
      security: [{ DeviceTypeHeader: [] }, { LanguageHeader: [] }, { ClientTokenHeader: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/ReferenceAdminResponse" }
  /v1/reference/dispatchers:
    get:
      tags: ["Reference / Freelance Dispatchers"]
      summary: "Справочник Freelance Dispatchers — work_status"
      security: [{ DeviceTypeHeader: [] }, { LanguageHeader: [] }, { ClientTokenHeader: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/ReferenceDispatchersResponse" }
...
  /v1/dispatchers/drivers:
    get:
      tags: ["Freelance Dispatchers / Приглашения водителей", "Freelance Dispatchers / My drivers"]
      summary: "Список моих водителей (с фильтрами и пагинацией)"
      description: |
        **Фриланс-диспетчер:** водители, привязанные к нему (freelancer_id = me). По этому списку диспетчер может отправлять им **таклиф** (рекомендацию груза) — POST /v1/dispatchers/cargo/:id/recommend (body: driver_id).
        **Фильтры:** phone (поиск по номеру), work_status (available, busy, loaded…), truck_type (тип тягача: TENT, REFRIGERATOR… из power_plate_type). **Пагинация:** page (с 1), limit (по умолчанию 20, макс. 100). **Сортировка:** sort=updated_at:desc (по умолчанию), name:asc, last_online_at:desc, work_status:asc.
        X-User-Token (dispatcher). Ответы на 5 языках (заголовок X-Language: ru, uz, en, tr, zh).
      security: [{ DeviceTypeHeader: [] }, { LanguageHeader: [] }, { ClientTokenHeader: [] }, { UserTokenHeader: [] }]
      parameters:
        - name: phone
          in: query
          schema: { type: string }
          description: "Поиск по номеру телефона (частичное совпадение)"
        - name: work_status
          in: query
          schema: { type: string, example: "available" }
          description: "Фильтр по статусу водителя (available, busy, loaded и т.д.)"
        - name: truck_type
          in: query
          schema: { type: string, example: "TENT" }
          description: "Фильтр по типу тягача (power_plate_type): TENT, REFRIGERATOR, FLATBED, TANKER, OTHER"
        - name: page
          in: query
          schema: { type: integer, default: 1, minimum: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
        - name: sort
          in: query
          schema: { type: string, default: "updated_at:desc", example: "updated_at:desc" }
          description: "Сортировка: updated_at:desc, name:asc, last_online_at:desc, work_status:asc"
      responses:
        "200": { description: "data.items — массив водителей, data.total — всего записей" }
...
  /v1/chat/ws:
    get:
      tags: [Chat]
      summary: "WebSocket чат"
      description: |
        Подключение к чату в реальном времени. Авторизация через query: **user_id=UUID** (для теста) или **token=JWT**.
        После подключения: отправка типа "typing" с data.conversation_id; получение событий "message" и "typing".

        **Дополнительно (Calls / WebRTC signaling):**
        Сервер принимает от клиента события `webrtc.offer`, `webrtc.answer`, `webrtc.ice`, `call.end` и пересылает peer'у, если отправитель является участником `call_id`.

        **Формат клиента → сервер:**
        `{ "type": "webrtc.offer", "data": { "call_id": "<uuid>", "payload": { ... } } }`

        **Формат сервер → peer:**
        `{ "type": "webrtc.offer", "data": { "call_id": "<uuid>", "from_id": "<uuid>", "payload": { ... } } }`
      security: [{ DeviceTypeHeader: [] }, { LanguageHeader: [] }, { ClientTokenHeader: [] }]
      parameters:
        - name: user_id
          in: query
          description: UUID пользователя (для теста без JWT)
          schema: { type: string, format: uuid }
        - name: token
          in: query
          description: JWT access token
          schema: { type: string }
      responses:
        "101": { description: Switching Protocols (WebSocket) }
        "401": { description: unauthorized }
...